+++ src/web/style/Workspaces.css
@@ -5,7 +5,7 @@ body {
   line-height: 1.5;
 }
 
-.Workspaces {
+.Workspaces__container {
   min-height: 100vh;
 }
